// src/Api/models/update-family-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sunday School API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { ObjectId } from './object-id';
 /**
 * 
 *
 * @export
 * @interface UpdateFamilyDto
 */
export interface UpdateFamilyDto {

    /**
     * @type {string}
     * @memberof UpdateFamilyDto
     */
    familyLastName?: string;

    /**
     * @type {string}
     * @memberof UpdateFamilyDto
     */
    familyAddress?: string;

    /**
     * @type {Array<ObjectId>}
     * @memberof UpdateFamilyDto
     */
    individuals?: Array<ObjectId>;
}